Move SmellConfiguration to top level.
