+++ lib/reek/smell_detectors/long_parameter_list.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-require_relative 'smell_configuration'
 require_relative 'base_detector'
 
 module Reek
